refactor(cli): share registry source-path resolution between fetch and callers
Move `resolveSourcePathFromIndex` out of `fetch.ts` into its own module so the
fetcher is not the only thing that can answer "which monorepo source backs this
install path at ref X?".

The update pruner (#1000) needs exactly that answer, and a second copy would be
free to drift: a pruner that disagreed with the fetcher about ownership could
delete a file the fetcher still installs. Behaviour is unchanged.

Refs #1000
